requirements.txt
requests~=2.32.5
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib